+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add `config.disable_session` to be able to disable session creation with specified
   authentication methods
+- Change `Cache-Control="no-store"` instead of `Cache-Control="no-cache, no-store"`,
+  and only set `Pragme="no-cache"` with HTTP 1.0 (and below).
 
 
 ## v1.8.3